requirements.txt
django
django-cors-headers
django-oauth-toolkit
djangorestframework
djangorestframework-oauth
